routing added, icons modified to use React Library
