+++ frontend/src/components/Navbar/Navbar.jsx
@@ -1,4 +1,10 @@
 import React from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import {
+  faSignInAlt,
+  faHome,
+  faUserCircle,
+} from '@fortawesome/free-solid-svg-icons';
 
 // Credit to: https://codepen.io/JFarrow/pen/fFrpg for the navbar
 
@@ -7,22 +13,27 @@ export default function Navbar() {
     <nav className="main-menu">
       <ul>
         <li>
-          <a href="">
-            <i className="fa fa-home fa-2x"></i>
+          <a href="/">
+            <div className="fa-home">
+              <FontAwesomeIcon icon={faHome} />
+            </div>
             <span className="nav-text">Dashboard</span>
           </a>
         </li>
         <li>
-          <a href="#">
-            <i className="fa fa-bar-chart-o fa-2x"></i>
-            <span className="nav-text">Graphs and Statistics</span>
+          <a href="/">
+            <div className="fa-home">
+              <FontAwesomeIcon icon={faSignInAlt} />
+            </div>
+            <span className="nav-text">Signup/Signin</span>
           </a>
         </li>
-
         <li>
-          <a href="#">
-            <i className="fa fa-table fa-2x"></i>
-            <span className="nav-text">Tables</span>
+          <a href="/">
+            <div className="fa-home">
+              <FontAwesomeIcon icon={faUserCircle} />
+            </div>
+            <span className="nav-text">Profile</span>
           </a>
         </li>
       </ul>
