+++ frontend/src/components/routes/SearchWalletBalance/_searchWalletBalance.scss
@@ -1,0 +1,38 @@
+// Need to refactor potentialliy
+.address-input-form {
+  // background-color: red;
+  height: 20vh;
+  flex-shrink: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.token-list {
+  width: 100%;
+  height: fit-content;
+  // background-color: aqua;
+  flex: 9;
+}
+
+// Buttons and Inputs
+input[type='text'],
+input[type='input'],
+input[type='password'] {
+  appearance: none;
+  background-color: none;
+  height: 60px;
+  width: 60%;
+  height: 30px;
+  padding: 10px 15px;
+  border-radius: $radius;
+  letter-spacing: 0.01em;
+}
+
+input[type='text']:focus,
+input[type='input']:focus,
+input[type='password']:focus,
+textarea:focus {
+  // border: none;
+  outline: none;
+}
