+++ frontend/src/styles/_imports.scss
@@ -7,3 +7,6 @@
 @import '../components/Navbar/navbar';
 @import '../components/dataTable/dataTable';
 @import '../components/cards/cryptoTokenBalance/cryptoTokenBalance';
+
+// Routes
+@import '../components/routes/SearchWalletBalance/searchWalletBalance';
